Add files via upload
Rational of the pipeline. The pipeline uses The Bacterial Transcription Factor Database (bacTFDB) and DeepARG approach to train (1. Training) a Deep Learning model named PredicTF.  PredicTF can use Genomes (a.1) or Metagenomes (b.1) as input, providing predictions of novel transcription factors that are listed in a file with their respective TF families (2. Prediction & Annotation). Finally, TFs listed and annotated can be mapped in the Transcriptomes (a.2) or Metatranscriptomes (b.2) providing a list of actives TFs in a specific condition (3. Mapping transcripts TFs).